Move alert service to services directory
